added utils folder and server membership checks in server and announcemnet apis

+++ backend/src/controllers/announcement.controller.js
@@ -1,28 +1,8 @@
 const Announcement = require("../models/announcement.model");
-const jwt = require("jsonwebtoken");
-const { Membership } = require("../models/server.model");
+const { checkUserInServer, checkUserIsAdmin } = require("../utils/util");
 
 const logEndPoint = (type, url) => {
     console.log(new Date().toLocaleString(), "--->", type, " ", url);
-};
-
-// check if user is member of server
-const checkUserInServer = async (serverId, userId) => {
-    const isMember = await Membership.exists({
-        server: serverId,
-        member: userId,
-    });
-    return isMember;
-};
-
-// check if user is admin of server
-const checkUserIsAdmin = async (serverId, userId) => {
-    const isAdmin = await Membership.exists({
-        server: serverId,
-        member: userId,
-        isAdmin: true,
-    });
-    return isAdmin;
 };
 
 // Create
@@ -84,11 +64,8 @@ exports.getAnnouncements = async (req, res) => {
         const userId = decoded.id;
         const serverId = req.body.serverId; // Assuming serverId is in the request parameters
 
-        console.log(userId);
-        console.log(serverId);
         // Check if the user is a member in the server
         const isMember = await checkUserInServer(serverId, userId);
-        console.log(isMember);
 
         if (isMember) {
             // User is a member and an admin, proceed to the next middleware or route
@@ -113,30 +90,55 @@ exports.getAnnouncements = async (req, res) => {
 };
 
 // Update
-exports.updateAnnouncement = async (req, res) => {
-    logEndPoint("PUT", "/announcements/:id");
+// exports.updateAnnouncement = async (req, res) => {
+//     logEndPoint("PUT", "/announcements/:id");
 
-    try {
-        // user can update title or description or both
-        const { title, description } = req.body;
-        await Announcement.findByIdAndUpdate(req.params.id, {
-            title,
-            description,
-        });
-        res.json({ message: "Announcement updated successfully" });
-    } catch (err) {
-        res.status(500).json({ message: err });
-    }
-};
+//     try {
+//         // user can update title or description or both
+//         const { title, description } = req.body;
+//         await Announcement.findByIdAndUpdate(req.params.id, {
+//             title,
+//             description,
+//         });
+//         res.json({ message: "Announcement updated successfully" });
+//     } catch (err) {
+//         res.status(500).json({ message: err });
+//     }
+// };
 
 // Delete
 exports.deleteAnnouncement = async (req, res) => {
     logEndPoint("DELETE", "/announcements/:id");
 
     try {
-        // hard delete right now
-        await Announcement.findByIdAndDelete(req.params.id);
-        res.json({ message: "Announcement deleted successfully" });
+        const decoded = req.decoded;
+        const userId = decoded.id;
+
+        const announcementId = req.params.id;
+        console.log(announcementId);
+        const announcement = await Announcement.findById(announcementId);
+        console.log("jjjj", announcement);
+        if (!announcement) {
+            return res.status(404).json({ error: "Announcement not found" });
+        }
+        const serverId = announcement.server._id;
+        console.log(serverId);
+
+        const isMember = await checkUserInServer(serverId, userId);
+        const isAdmin = await checkUserIsAdmin(serverId, userId);
+        console.log(isMember);
+        console.log(isAdmin);
+
+        if (isMember && isAdmin) {
+            // hard delete right now
+            await Announcement.findByIdAndDelete(req.params.id);
+            res.json({ message: "Announcement deleted successfully" });
+        } else {
+            // User is not a member or not an admin, return forbidden status
+            return res
+                .status(403)
+                .json({ message: "Forbidden, user not authorized" });
+        }
     } catch (err) {
         res.status(500).json({ message: err });
     }
